diseases feature matrix, dimension 15455 x 10682
R data.frame of dimension 15455 x 10682 containing 15455 ICD9 codes in the first column and in the remaining 10681 columns features extracted from biomedical databases. The element [i,j] of the matrix is 1, if the i-th ICD9-Code is associated with the j-th feature and 0 otherwise.

Features names are grouped to so called "domains" regarding to it's origin / context.
A particular column has the naming <domain>.x.<feature>, e. g. "DISEASE.PHEWAS.x.MULTIPLE SCLEROSIS" or "DISEASE.BIOMARKER.x.ALDOSTERONE".

The following domains are contained:
1) PheWAS: <DISEASE.PHEWAS>
2) PheWAS highlevel (aggregation of PheWAS from format xxx.xx to xxx): <DISEASE.PHEWAS_HL>
3) GO (biological process only): <DISEASE.GO>
4) Pathways (KEGG and WikiPathways): <DISEASE.PATH>
5) MeSH Disease: <DISEASE.MESHDISEASE>
6) Symptoms: <DISEASE.SYMPTOMS>
7) Biomarker: <DISEASE.BIOMARKER>