drugfeature matrix, dimension 1088 x 7591
R data.frame of dimension 1008 x 7591 containing 1008 substancenames in the first column and in the remaining 7590 columns features extracted from biomedical databases. The element [i,j] of the matrix is 1, if the i-th substancename is associated with the j-th feature and 0 otherwise.

Features names are grouped to so called "domains" regarding to it's origin / context.
A particular column has the naming <domain>.x.<feature>, e. g. "DRUG.PHEWAS.x.HYPERTENSIVE CHRONIC KIDNEY DISEASE" or "DRUG2SIDEEFFECTxFREQ.x.HYPOKALAEMIA.x.COMMON". In the latter case, prescribing the potential drug sideeffects and their frequencies, the naming is extended by a frequency term: <domain>.x.<feature>.x.<frequency>.

The following domains are contained:
1) Therapy Class (from Redbook): <DRUG.THRCLDS>
2) Therapy Group: >DRUG.THRGRDS>
3) Drug Indication in terms of PheWAS: <DRUG.PHEWAS>
4) Drug Indication in terms of PheWAS highlevel: <DRUG.PHEWAS_HL>
5) GO-Terms (biological proc